requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.6.0
tensorboard>=2.11.0
